unsigned long long frequency

+++ Huffman/ByteFrequency.h
@@ -11,20 +11,20 @@ using namespace std;
 class ByteFrequency
 {
     private:
-        unsigned int frequency;
+        unsigned long long frequency;
         byte byte_code;
 
     public:
-        unsigned int get_frequency() const;
+        unsigned long long get_frequency() const;
         byte get_byte_code() const;
 
-        void set_frequency(unsigned int f);
+        void set_frequency(unsigned long long f);
         void set_byte_code(byte b);
         void increment_frequency();
 
         bool is_empty();
 
-        ByteFrequency(unsigned int f, byte b);
+        ByteFrequency(unsigned long long f, byte b);
         ByteFrequency();
 
         friend ofstream& operator<< (ofstream& os, const ByteFrequency& bf);
